DocumentSimilarity/requirements.txt
flask
scikit-learn
numpy
python-docx
PyPDF2